fix(symbolic): skip corpus seeds for gas-dependent paths
Refusing gasleft constraints in `model()` also aborted the seed
materialization that runs on successful paths when fuzz corpus seeding
is enabled, downgrading a proved-safe run to incomplete. Skip the seed
for such a path instead, count model queries only after the guard, keep
the Echidna flags test on the default check interval so shortest-prefix
minimization stays covered, move the `SymBytes` test next to the type,
narrow the README gas row to what the engine actually does, and pin the
exact z3 wheel version.
