+++ .github/workflows/test.yml
@@ -102,7 +102,7 @@ jobs:
       - name: Install Z3
         # Pinned so the `forge test --symbolic` snapshot tests stay stable across solver versions.
         if: runner.os == 'Linux'
-        run: pip install z3-solver==5.1.0 && z3 --version
+        run: pip install z3-solver==5.1.0.0 && z3 --version
 
       - name: Foundry test cache
         uses: actions/cache@55cc8345863c7cc4c66a329aec7e433d2d1c52a9 # v6.1.0
